All dependencies reconfigured to compile from source

+++ RigelRenderer/source/internal/utility/FreeTypeUtility.cpp
@@ -62,7 +62,7 @@ namespace rgr
                 texture,
                 glm::ivec2(face->glyph->bitmap.width, face->glyph->bitmap.rows),
                 glm::ivec2(face->glyph->bitmap_left, face->glyph->bitmap_top),
-                face->glyph->advance.x
+                static_cast<unsigned int>(face->glyph->advance.x)
             };
             texturesMapOutput->insert(std::pair<char, Character>(c, character));
         }
